Replaced scrypt with pylibscrypt, so there's a pure-python fallback on C-less platforms
As I'd like to get deadlock running on platforms that are poorly supportive
of native compilation, such as Android and other embedded platforms, I'd like
pure-python fallbacks for the core crypto in deadlock. Yes, naturally this means
very slow crypto on these platforms, but better than none!

To this end this commit replaces the "scrypt" dependency with pylibscrypt.
For blake2 and nacl, native versions may be bundled and setup.py tweaked
not to fail-out if it can't install C versions (but to try, at least).